requirements.txt
Django==4.2.5
djangorestframework==3.14.0
mysqlclient>=2.0
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5
celery>=5.0
django-celery-beat>=2
pylint==2.9.6
pylint-celery==0.3      
pylint-django==2.4.4    
pylint-plugin-utils==0.6
redis>=3.5
requests==2.26.0
dateparser==1.0.0
simplejson==3.17.3
pandas==1.3.0
fake-useragent==1.1.1
flatten-dict==0.4.2
lxml==4.6.3
SQLAlchemy==1.4.23
PyMySQL==1.0.2
beautifulsoup4==4.8.2
#bs4==0.0.1
gunicorn==20.1.0
whitenoise==6.2.0
djangorestframework-datatables==0.7.0
boto==2.49.0
boto3==1.19.9
django-storages==1.12.3
django-filter==22.1
nltk==3.6.2
pdf2image==1.16.0
#poppler-utils==0.1.0
PyPDF2==1.26.0
pytesseract==0.3.8
regex==2021.8.3
Filetype==1.0.7
opencv-python==4.4.0.46
PyMuPDF==1.18.9
Pillow==8.3.1
slack-sdk==3.21.3
django-cors-headers==4.2.0
drf-writable-nested==0.7.0
wget==3.2
django-extensions==3.1.3
selenium==4.20.0
webdriver-manager==3.4.2
numpy==1.26.4
openpyxl==3.1.5
python-dateutil==2.8.2 
#celery==5.2.7 
django-registration-redux==2.9
weasyprint==53.0 
sentry-sdk==2.22.0
python-dotenv==1.0.1

